[NB] Set default UI corners on GUI startup
Wait for the GUI process to load and connect to the 0MQ plugin hub.
After GUI process plugin has connected to the hub, restore the default
canvas and (video) frame corners from the corresponding app options.

The corners affect the registration of the video to the device drawing
(i.e., electrodes).