fix(process): only kill process if running
Prior to this commit, an attempt was made to kill device UI process even
if the process was no longer running, e.g., if the user manually closed
the UI window.  While this behaviour did not cause any lasting issue, it
resulted in an error message logged to the console, which may
confuse/concern the user.

In this commit, only kill the device UI process if it is running.