chore(zmq-hub): use background thread hub_execute
Prior to this commit, `microdrop.plugin_helpers.hub_execute()` was assumed
to be running in the MicroDrop main GTK thread.  To prevent blocking the
UI, this plugin executed `refresh_gui()` to update the UI while waiting
for `hub_execute()` to finish.  However, as of MicroDrop 2.25, the
`hub_execute()` function is run in a separate thread.  Running
`refresh_gui()` in this thread would violate GTK thread safety.

In this commit:

 - Require `microdrop >=2.25`; and
 - _Do not_ execute `refresh_gui()` while waiting for a response from
   `hub_execute()`.